Project Work - Early work (v0.1)
Beginnings of the Project work file.  I have included a new working
horizontal bar (little bit more stylized and functional, with some
pull-down).  Working to have a library of songs with song details that
are seperate (but related) to a given user project.   Songs and the
phrases will not change over time, but the projects will (as they assign
various models to songs and effects to phrases).