Project v1.08
Added Username to the Song structure.  If the song is owned by  user 'f'
then this will be shown as a 'global' song (accessible by all).  If the
song is owned by other than 'f' it is a 'local' song and only viewable
by the owner.

In projects, a user can sequence to songs that are either global or
local (aka owned by themselves or user 'f')

Kurt