Latest changes
Some of this is not finished.
Fixed Song_add to have a MAXFILESIZE field
Modded Songs.php to include the error checking and items from gloabl
library that do not have phrasing.
gen_xml.php is a library that will be used in the future for export and
import of projects as XML.
songglobal_add.php is the beginning (not complete) of the code to handle
the copying of global songs to local and allow users to attach phrase
files to global songs if they don't exist already.