Fixed some local errors
First - Made sure to get rid of the attempts to translate arrays values
as strings.  With the verbose error checking the program didn't like
this.
Second - gnuplot wasn't point to the right place (probably worked with
your local because gnuplot was in your path).  Did an absolute path to
gnuplot executable using the $_SERVER command to find root and go from
there if we are running on localhost.