Project v1.10
Major upgrades to project and project_filer to use effect_id instead of
username/effect pair in project_detail.  NOTE:  This requires a rebuild
of the project data, since there is an added field 'effect_id' in
project_detail.  Without this change, these updates to project will not
work.
ADDED FUNCTION : Now a user can click on the thumbnail of an effect and
edit that effect directly from project.  The changes to the parameters
of the effect are saved (if the user clicks submit) and they are
returned to the phrase edit window of project after the submit or
cancel.